+++ src/com/uet/int2204/group2/entity/StaticEntity.java
@@ -2,6 +2,7 @@ package com.uet.int2204.group2.entity;
 
 import com.uet.int2204.group2.utils.Constants;
 
+// static entities are tied to a tile
 public abstract class StaticEntity extends Entity {
   private int tileX;
   private int tileY;
@@ -11,11 +12,11 @@ public abstract class StaticEntity extends Entity {
     this.tileY = y;
   }
 
-  @Override public int getPixelX() {
+  @Override public double getPixelX() {
     return getTileX() * Constants.TILE_SIZE;
   }
 
-  @Override public int getPixelY() {
+  @Override public double getPixelY() {
     return getTileY() * Constants.TILE_SIZE;
   }
